requirements.txt
rknn-toolkit2